Add weighted round robin scheduer for partitions
We add this scheduler, even the FP, EDF and RR scheduler for partitions based on original configurations. It means that running slots for partitions are configured by the #define statement. But we need to add deadline and weight to the pok_partition_t struct. How could a program pass these arguments to the kernel. On the one hand, we could use the #define statement as before we define the running slots. On the other hand, we could add a new syscall to add a partition dynamically as the pok_thread_create function. In this way, we may need to modify more code but definitely the code would be more pretty.